Replace java_toolchain with default_java_toolchain.
Bazel is changing how Java toolchains are resolve. I updated the
instructions and replaced direct call to java_toolchain with
default_java_toolchain macro.

The macro is easier to maintain, because if java_toolchain rule is changed it is
easier to fix it in the macro without the need to update all downstream
repositories as well.